--- learning/tour-of-beam/learning-content/java/schema-based-transforms/join/example/Task.java
@@ -1,182 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
-/*
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// beam-playground:
-//   name: join
-//   description: Join example.
-//   multifile: false
-//   context_line: 46
-//   categories:
-//     - Quickstart
-//   complexity: ADVANCED
-//   tags:
-//     - hellobeam
-
-import lombok.EqualsAndHashCode;
-import org.apache.beam.sdk.Pipeline;
-import org.apache.beam.sdk.io.TextIO;
-import org.apache.beam.sdk.options.PipelineOptions;
-import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.schemas.JavaFieldSchema;
-import org.apache.beam.sdk.schemas.Schema;
-import org.apache.beam.sdk.schemas.annotations.DefaultSchema;
-import org.apache.beam.sdk.schemas.annotations.SchemaCreate;
-import org.apache.beam.sdk.schemas.transforms.Convert;
-import org.apache.beam.sdk.schemas.transforms.Group;
-import org.apache.beam.sdk.schemas.transforms.Join;
-import org.apache.beam.sdk.schemas.transforms.RenameFields;
-import org.apache.beam.sdk.transforms.*;
-import org.apache.beam.sdk.values.PCollection;
-import org.apache.beam.sdk.values.Row;
-import org.apache.beam.sdk.values.TypeDescriptor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-
-public class Test {
-    private static final Logger LOG = LoggerFactory.getLogger(Test.class);
-
-    @DefaultSchema(JavaFieldSchema.class)
-    @EqualsAndHashCode
-    public static class Game {
-        public String userId;
-        public Integer score;
-        public String gameId;
-        public String date;
-
-        @SchemaCreate
-        public Game(String userId, Integer score, String gameId, String date) {
-            this.userId = userId;
-            this.score = score;
-            this.gameId = gameId;
-            this.date = date;
-        }
-
-        @Override
-        public String toString() {
-            return "Game{" +
-                    "userId='" + userId + '\'' +
-                    ", score='" + score + '\'' +
-                    ", gameId='" + gameId + '\'' +
-                    ", date='" + date + '\'' +
-                    '}';
-        }
-    }
-
-    // User schema
-    @DefaultSchema(JavaFieldSchema.class)
-    @EqualsAndHashCode
-    public static class User {
-
-        public String userId;
-        public String userName;
-
-        @SchemaCreate
-        public User(String userId, String userName) {
-            this.userId = userId;
-            this.userName = userName;
-        }
-
-        @Override
-        public String toString() {
-            return "User{" +
-                    "userId='" + userId + '\'' +
-                    ", userName='" + userName + '\'' +
-                    '}';
-        }
-    }
-
-
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.fromArgs(args).create();
-        Pipeline pipeline = Pipeline.create(options);
-
-        PCollection<User> userInfo = getUserPCollection(pipeline);
-        PCollection<Game> gameInfo = getGamePCollection(pipeline);
-
-        Schema type = Schema.builder()
-                .addStringField("userId")
-                .addStringField("userName")
-                .build();
-
-        PCollection<Row> pCollection = userInfo
-                .apply(MapElements.into(TypeDescriptor.of(Object.class)).via(it -> it))
-                .setSchema(type,
-                        TypeDescriptor.of(Object.class), input ->
-                        {
-                            User user = (User) input;
-                            return Row.withSchema(type)
-                                    .addValues(user.userId, user.userName)
-                                    .build();
-                        },
-                        input -> new User(input.getString(0), input.getString(1))
-                )
-                .apply(Join.innerJoin(gameInfo).using("userId"));
-
-        pCollection
-                .apply("User flatten row", ParDo.of(new LogOutput<>("Flattened")));
-
-        pipeline.run();
-    }
-
-    public static PCollection<User> getUserPCollection(Pipeline pipeline) {
-        PCollection<String> rides = pipeline.apply(TextIO.read().from("gs://apache-beam-samples/game/small/gaming_data.csv"));
-        final PTransform<PCollection<String>, PCollection<Iterable<String>>> sample = Sample.fixedSizeGlobally(100);
-        return rides.apply(sample).apply(Flatten.iterables()).apply(ParDo.of(new ExtractUserFn()));
-    }
-
-    public static PCollection<Game> getGamePCollection(Pipeline pipeline) {
-        PCollection<String> rides = pipeline.apply(TextIO.read().from("gs://apache-beam-samples/game/small/gaming_data.csv"));
-        final PTransform<PCollection<String>, PCollection<Iterable<String>>> sample = Sample.fixedSizeGlobally(100);
-        return rides.apply(sample).apply(Flatten.iterables()).apply(ParDo.of(new ExtractUserProgressFn()));
-    }
-
-    static class ExtractUserFn extends DoFn<String, User> {
-        @ProcessElement
-        public void processElement(ProcessContext c) {
-            String[] items = c.element().split(",");
-            c.output(new User(items[0], items[1]));
-        }
-    }
-
-    static class ExtractUserProgressFn extends DoFn<String, Game> {
-        @ProcessElement
-        public void processElement(ProcessContext c) {
-            String[] items = c.element().split(",");
-            c.output(new Game(items[0], Integer.valueOf(items[2]), items[3], items[4]));
-        }
-    }
-
-    static class LogOutput<T> extends DoFn<T, T> {
-
-        private final String prefix;
-
-        LogOutput() {
-            this.prefix = "Processing element";
-        }
-
-        LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            LOG.info(prefix + ": {}", c.element());
-        }
-    }
-}
